rename component back to Timer rel: #11
